homework final

+++ src/main/java/de/danilova/mymarket/specifications/ProductSpecification.java
@@ -3,14 +3,15 @@ package de.danilova.mymarket.specifications;
 import de.danilova.mymarket.models.Product;
 import org.springframework.data.jpa.domain.Specification;
 
+import java.math.BigDecimal;
 
 
 public class ProductSpecification {
-    public static Specification<Product> priceGreaterOrEqualsThen(Long minPrice){
+    public static Specification<Product> priceGreaterOrEqualsThen(BigDecimal minPrice){
         return ((root, query, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("price"),minPrice));
     }
 
-    public static Specification<Product> priceLesserOrEqualsThen(Long maxPrice){
+    public static Specification<Product> priceLesserOrEqualsThen(BigDecimal maxPrice){
         return ((root, query, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("price"),maxPrice));
     }
